refactor(web-ui): isolate high-frequency braille animation state
Extracted the 100ms `setInterval` state used for the braille halo spinner from `LandingPage` and `WelcomeScreen` into its own isolated `HaloSpinner` component wrapped in `React.memo`. This prevents full tree re-renders ten times a second.
