+++ web-ui/src/components/Chat/WelcomeScreen.tsx
@@ -1,17 +1,6 @@
-import { useEffect, useState } from 'react';
-import { SPINNER_FRAMES } from '../../constants/spinner';
+import { HaloSpinner } from '../ui/HaloSpinner';
 
 export function WelcomeScreen() {
-  const [brailleOffset, setBrailleOffset] = useState(0);
-
-  // Animate braille halo when welcome screen is visible
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setBrailleOffset(prev => (prev + 1) % SPINNER_FRAMES.length);
-    }, 100);
-    return () => clearInterval(interval);
-  }, []);
-
   return (
     <div className="relative flex items-center justify-center h-full px-6 bg-bg-100 overflow-hidden">
       {/* Background watermark layer */}
@@ -21,25 +10,7 @@ export function WelcomeScreen() {
           OpenDev
         </span>
         {/* Orbiting braille halo ring */}
-        <div className="absolute animate-spin-slow" style={{ width: 360, height: 360 }}>
-          {Array.from({ length: 24 }).map((_, i) => {
-            const angle = (i / 24) * 360;
-            const char = SPINNER_FRAMES[(i + brailleOffset) % SPINNER_FRAMES.length];
-            return (
-              <span
-                key={i}
-                className="absolute text-lg font-mono text-bg-300"
-                style={{
-                  left: '50%',
-                  top: '50%',
-                  transform: `rotate(${angle}deg) translateX(180px) rotate(-${angle}deg)`,
-                }}
-              >
-                {char}
-              </span>
-            );
-          })}
-        </div>
+        <HaloSpinner />
       </div>
       {/* Foreground welcome content */}
       <div className="relative z-10 text-center">
